refactor(fe): add admin/_libs folder
