+++ apps/frontend/app/admin/problem/_components/LimitForm.tsx
@@ -4,7 +4,7 @@ import { Input } from '@/components/ui/input'
 import { cn } from '@/lib/utils'
 import { useFormContext } from 'react-hook-form'
 import ErrorMessage from '../../_components/ErrorMessage'
-import { inputStyle } from '../../utils'
+import { inputStyle } from '../../_libs/utils'
 
 export default function LimitForm({ blockEdit }: { blockEdit?: boolean }) {
   const {
